rag-cv/requirements.txt
# RAG CV Dependencies
# ===================
# Production dependencies for the RAG CV Analysis system

# Core Framework
streamlit>=1.28.0

# LangChain Ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# Optional: For semantic chunking
langchain-experimental>=0.0.60

# Embeddings
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0

# Vector Database
chromadb>=0.4.0

# PDF Processing
pypdf>=3.0.0

# Optional: Better PDF extraction
# pymupdf>=1.23.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# Testing
pytest>=7.0.0

# Utilities
tiktoken>=0.5.0
langchain-ollama
